Modifies Color to store R,G,B and also to do addition and division for averaging.
Works on pickup sequence:

Turntable settings should work, but colors were not reading perfectly (only identified red and yellow so far).

So storing coins works perfectly fine, its the identification that does not

TODO: Write code for randomly assigning an unknown coin to a cup. It is better to misplace a coin in a square than to drop it